Consolidate conftest in root
